Windows implementation of privilege escalation.
The built in 'runas' command was found to be nowhere near sufficient, so 
A GUI helper application has been created that acts like GkSudo on
Linux. 

This uses JNA to access the CreateProcessWithLoginW function (and all
the other stuff you have to do such as connecting pipes and obtaining
environments).

Username, domain and password may be passed to it along with an
application to run and its arguments. The password will be prompted for
if it is not supplied, and std i/o streams are connected.

While this should all be sufficient for loading the daemon as an
administrator, and ad-hoc escalation (or running as another user), it
probably won't work for the case of an administrator daemon spawning a
user PTY. Another solution will be needed for that.