If the forker daemon is in use, the 'classpath' does not have a '+', and the forker utilities library is not included in that classpath, the forker daemon will not be found.
This regression was introduced by the +/- feature that was intended for *boot* classpath. I ended up re-using the same code, but didn't consider the backward compatibility issue that might arise from doing so.

No the lack or + or - on the classpath, means append is used by default. If you WANT to replace the classpath entirely, the path should now be prefixed with a =